rename Table component
