+++ packages/gp2-components/src/molecules/index.ts
@@ -1,4 +1,6 @@
 export { default as CardWithBackground } from './CardWithBackground';
+export { default as CollapsibleTable } from './CollapsibleTable';
+export { default as ContactSupport } from './ContactSupport';
 export { default as ExpandableText } from './ExpandableText';
 export { default as HeaderLogo } from './HeaderLogo';
 export { default as IconWithLabel } from './IconWithLabel';
@@ -8,5 +10,3 @@ export { default as UserCardInfo } from './UserCardInfo';
 export { default as UserMenu } from './UserMenu';
 export { default as UserPositions } from './UserPositions';
 export { default as UserRegion } from './UserRegion';
-export { default as ContactSupport } from './ContactSupport';
-export { default as TableDisplay } from './TableDisplay';
